ai_service/ai_integration.py
"""
AI Integration Module
Handles communication with AI APIs (ChatGPT, Claude, etc.)
"""

import openai
import anthropic
import os
import json
from typing import Optional, Dict, Any
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

class AIIntegration:
    """Handles AI API integration with Christian focus"""
    
    def __init__(self):
        # TODO: Initialize API clients
        # TODO: Set current provider
        # TODO: Set up Christian context
        # TODO: Initialize clients
        pass
    
    def _initialize_clients(self):
        """Initialize API clients"""
        # TODO: Initialize OpenAI client
        # TODO: Initialize Anthropic client
        # TODO: Handle initialization errors
        pass
    
    def set_provider(self, provider: str):
        """Set the AI provider to use"""
        # TODO: Validate provider
        # TODO: Set current provider
        pass
    
    def format_christian_prompt(self, user_input: str) -> str:
        """Format user input with Christian context"""
        # TODO: Combine Christian context with user input
        # TODO: Return formatted prompt
        pass
    
    def send_message_to_ai(self, message: str) -> Optional[str]:
        """Send message to AI and get response"""
        # TODO: Route to appropriate provider
        # TODO: Handle errors
        # TODO: Return AI response
        pass
    
    def _send_to_openai(self, message: str) -> Optional[str]:
        """Send message to OpenAI ChatGPT"""
        # TODO: Format prompt
        # TODO: Send to OpenAI API
        # TODO: Handle API errors
        # TODO: Return response
        pass
    
    def _send_to_anthropic(self, message: str) -> Optional[str]:
        """Send message to Anthropic Claude"""
        # TODO: Format prompt
        # TODO: Send to Anthropic API
        # TODO: Handle API errors
        # TODO: Return response
        pass
    
    def process_ai_response(self, response: str) -> str:
        """Process and format AI response"""
        # TODO: Clean up response
        # TODO: Add Christian closing if appropriate
        # TODO: Return processed response
        pass
    
    def get_available_providers(self) -> Dict[str, bool]:
        """Get list of available AI providers"""
        # TODO: Check which providers are configured
        # TODO: Return availability status
        pass
    
    def test_connection(self) -> Dict[str, Any]:
        """Test connection to AI providers"""
        # TODO: Test OpenAI connection
        # TODO: Test Anthropic connection
        # TODO: Return test results
        pass 