requirements.txt
# ===== CORE FRAMEWORKS =====

# Voice Recognition
speech_recognition==3.10.0
pyaudio==0.2.11

# AI Integration
openai==1.3.0
anthropic==0.7.0
python-dotenv==1.0.0

# Text-to-Speech
pyttsx3==2.90

# Web Server (Flask)
flask==3.0.0
flask-cors==4.0.0

# System Utilities
psutil==5.9.6
requests==2.31.0

# ===== RASPBERRY PI SPECIFIC =====
# These may need system-level installation:
# sudo apt-get install espeak-ng
# sudo apt-get install portaudio19-dev
# sudo apt-get install python3-pyaudio

# ===== INSTALLATION =====
# pip install -r requirements.txt 